+++ basicswap/interface/dcr/dcr.py
@@ -14,7 +14,11 @@ import traceback
 
 from typing import List, Optional
 
-from basicswap.basicswap_util import getVoutByScriptPubKey, TxLockTypes
+from basicswap.basicswap_util import (
+    ADAPTOR_SIG_LOCK_SPEND_FEE_BUFFER,
+    getVoutByScriptPubKey,
+    TxLockTypes,
+)
 from basicswap.chainparams import Coins
 from basicswap.contrib.test_framework.script import (
     CScriptNum,
@@ -1231,7 +1235,14 @@ class DCRInterface(FeeValidator, Secp256k1Interface):
         return script
 
     def createSCLockSpendTx(
-        self, tx_lock_bytes, script_lock, pkh_dest, tx_fee_rate, vkbv=None, fee_info={}
+        self,
+        tx_lock_bytes,
+        script_lock,
+        pkh_dest,
+        tx_fee_rate,
+        vkbv=None,
+        fee_info={},
+        tx_lock_refund_bytes=None,
     ):
         tx_lock = self.loadTx(tx_lock_bytes)
         output_script = self.getScriptDest(script_lock)
@@ -1249,7 +1260,13 @@ class DCRInterface(FeeValidator, Secp256k1Interface):
         dummy_witness_stack = self.getScriptLockTxDummyWitness(script_lock)
         size = len(self.setTxSignature(tx.serialize(), dummy_witness_stack))
         size += 1
-        pay_fee = self.feeForVSize(tx_fee_rate, size)
+        if tx_lock_refund_bytes is None:
+            pay_fee = self.feeForVSize(tx_fee_rate, size)
+        else:
+            pay_fee = (
+                self.getLockRefundTxFee(locked_coin, tx_lock_refund_bytes)
+                + ADAPTOR_SIG_LOCK_SPEND_FEE_BUFFER
+            )
         tx.vout[0].value = locked_coin - pay_fee
 
         fee_info["fee_paid"] = pay_fee
@@ -1458,7 +1475,14 @@ class DCRInterface(FeeValidator, Secp256k1Interface):
         return txid, locked_n
 
     def verifySCLockSpendTx(
-        self, tx_bytes, lock_tx_bytes, lock_tx_script, a_pkhash_f, feerate, vkbv=None
+        self,
+        tx_bytes,
+        lock_tx_bytes,
+        lock_tx_script,
+        a_pkhash_f,
+        feerate,
+        vkbv=None,
+        tx_lock_refund_bytes=None,
     ):
         # Verify:
         #   Must have only one input with correct prevout (n is always 0) and sequence
@@ -1510,8 +1534,17 @@ class DCRInterface(FeeValidator, Secp256k1Interface):
             fee_rate_paid,
         )
 
-        if not self.compareFeeRates(fee_rate_paid, feerate):
-            raise ValueError(f"Bad fee rate, expected: {feerate}")
+        if tx_lock_refund_bytes is None:
+            if not self.compareFeeRates(fee_rate_paid, feerate):
+                raise ValueError(f"Bad fee rate, expected: {feerate}")
+            return True
+
+        expect_fee: int = (
+            self.getLockRefundTxFee(locked_coin, tx_lock_refund_bytes)
+            + ADAPTOR_SIG_LOCK_SPEND_FEE_BUFFER
+        )
+        if fee_paid != expect_fee:
+            raise ValueError(f"Bad fee, expected: {expect_fee}, paid: {fee_paid}")
 
         return True